+++ app/Http/Controllers/FournisseurController.php
@@ -6,6 +6,7 @@ use App\Models\Fournisseur;
 use Illuminate\Http\Request;
 use Illuminate\Database\QueryException;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Auth\Access\AuthorizationException;
 
 class FournisseurController extends Controller
 {
@@ -14,8 +15,15 @@ class FournisseurController extends Controller
      */
     public function index()
     {
-        $fournisseur = Fournisseur::all();
-        return response()->json(['fournisseur' => $fournisseur]);
+        try {
+            $this->authorize('view', Fournisseur::class);
+            $fournisseur = Fournisseur::all();
+            return response()->json(['message' => 'Liste des client récupérée avec succès', 'fournisseur' =>  $fournisseur], 200);
+        } catch (AuthorizationException $e) {
+            return response()->json(['error' => 'Vous n\'avez pas l\'autorisation de voir la liste des fournisseur.'], 403);
+        } catch (\Exception $e) {
+            return response()->json(['error' => $e->getMessage()], 500);
+        }
     }
 
     /**
@@ -31,23 +39,29 @@ class FournisseurController extends Controller
      */
     public function store(Request $request)
     {
-        // validation 
-        $validator = Validator::make(
-            $request->all(),
-            [
+        try {
+            $this->authorize('add', Fournisseur::class);
+
+            // Validation 
+            $validator = Validator::make($request->all(), [
                 'raison_sociale' => 'required',
                 'adresse' => 'required',
                 'tele' => 'required',
                 'ville' => 'required',
                 'abreviation' => 'required',
                 'zone' => 'required',
-            ]
-        );
-        if ($validator->fails()) {
-            return response()->json(['error' => $validator->errors()], 400);
-        } else {
+            ]);
+
+            if ($validator->fails()) {
+                return response()->json(['error' => $validator->errors()], 400);
+            }
+
             $fournisseur = Fournisseur::create($request->all());
-            return response()->json(['message' => 'Fournisseur ajouteé avec succès', 'fournisseur' => $fournisseur], 200);
+            return response()->json(['message' => 'fournisseur ajoutée avec succès', 'fournisseur' => $fournisseur], 200);
+        } catch (AuthorizationException $e) {
+            return response()->json(['error' => 'Vous n\'avez pas l\'autorisation de modifier cette fournisseur.'], 403);
+        } catch (\Exception $e) {
+            return response()->json(['error' => $e->getMessage()], 500);
         }
     }
 
@@ -73,24 +87,29 @@ class FournisseurController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $fournisseur = Fournisseur::findOrFail($id);
+        try {
+            $this->authorize('modify', Fournisseur::class);
+            $fournisseur = Fournisseur::findOrFail($id);
 
-        $validator = Validator::make(
-            $request->all(),
-            [
+            $validator = Validator::make($request->all(), [
                 'raison_sociale' => 'required',
                 'adresse' => 'required',
                 'tele' => 'required',
                 'ville' => 'required',
                 'abreviation' => 'required',
                 'zone' => 'required',
-            ]
-        );
-        if ($validator->fails()) {
-            return response()->json(['error' => $validator->errors()], 400);
-        } else {
+            ]);
+
+            if ($validator->fails()) {
+                return response()->json(['error' => $validator->errors()], 400);
+            }
+
             $fournisseur->update($request->all());
-            return response()->json(['message' => 'Fournisseur modifié avec succès', 'fournisseur' => $fournisseur], 200);
+            return response()->json(['message' => 'fournisseur modifié avec succès', 'fournisseur' => $fournisseur], 200);
+        } catch (AuthorizationException $e) {
+            return response()->json(['error' => 'Vous n\'avez pas l\'autorisation de modifier cette fournisseur.'], 403);
+        } catch (\Exception $e) {
+            return response()->json(['error' => $e->getMessage()], 500);
         }
     }
 
@@ -100,13 +119,18 @@ class FournisseurController extends Controller
     public function destroy($id)
     {
         try {
-
+            $this->authorize('delete', Fournisseur::class);
             $fournisseur = Fournisseur::findOrFail($id);
             $fournisseur->delete();
-            return response()->json(['message' => 'Le fournisseur a été supprimé avec succès.'], 200);
+
+            return response()->json(['message' => 'Fournisseur supprimée avec succès'], 200);
+        } catch (AuthorizationException $e) {
+            return response()->json(['error' => 'Vous n\'avez pas l\'autorisation de supprimer cette Fournisseur.'], 403);
+        } catch (\Exception $e) {
+            return response()->json(['error' => $e->getMessage()], 500);
         } catch (QueryException $e) {
-            // Si une exception est déclenchée, cela signifie que le fournisseur a des commandes associées
-            return response()->json(['error' => 'Impossible de supprimer ce fournisseur car il a des produits associées.'], 400);
+            // Si une exception est déclenchée, cela signifie que le client a des commandes associées
+            return response()->json(['error' => 'Impossible de supprimer ce Fournisseur car il a des produits associées.'], 400);
         }
     }
 }
